requirements.txt
boto3>=1.34.0
pytest>=8.0.0
matplotlib>=3.8.0

streamlit
pandas
matplotlib
numpy
pyyaml